pt: Consolidate page table related operations
move fundamental page table ops and definitions such as identity page table
mapping and global page size under the pt_lib library
