+++ include/hobos/mmu.h
@@ -117,7 +117,8 @@ struct tcr_el1_cfg {
 #define MAIR_MEM_I_WB_NT(r, w)	MAIR_MEM_RW(0b11, r, w)
 
 //kernel offset after switch to high memory addressing
-#define KERNEL_START	0xFFFFFF8000000000
+#define KERNEL_START	0xFFFFFF8000000000UL
+#define PAGE_SIZE	KB(4)
 
 struct mair_attr {
 	char outer_cache: 4;
@@ -138,10 +139,6 @@ extern u8 pt_ctr;
 
 void init_mmu(void);
 void switch_vmem(void);
-void map_pa_to_va_pg(unsigned long pa, unsigned long va,
-		     struct page_table_desc *pt_top,
-		     unsigned long flags);
-void create_id_mapping(u64 start_paddr, u64 end_paddr,
-			u64 pt, u64 flags);
+void set_ttbr0_el1(u64 x);
 
 #endif
