+++ pt.c
@@ -3,10 +3,12 @@
 #include <hobos/kstdio.h>
 #include <hobos/lib/pt_lib.h>
 #include <hobos/asm/barrier.h>
+#include <hobos/mmu.h>
+#include <hobos/compiler_types.h>
 
 //global list of page tables
 //TODO: make size dynamic
-struct page_table_desc *global_page_tables[10] __attribute__((section(".misc"))) ;
+__section(".misc") struct page_table_desc *global_page_tables[10];
 u8 pt_ctr;
 
 // for simplicity, we will assume that this OS only cares about
@@ -53,7 +55,7 @@ void place_pt_entry(struct page_table_desc *pt_desc, u64 pte, int index)
 //
 //returns: the first new page table entry created
 volatile u64 *create_pt_entries(struct page_table_desc *pt_desc,
-		       u64 start_paddr, u64 end_paddr,
+				u64 start_paddr, u64 end_paddr,
 		       u64 flags)
 {
 	volatile u64 *pt = pt_desc->pt;
@@ -107,6 +109,95 @@ struct page_table_desc *create_pt(u64 pt_baddr, char level)
 	global_page_tables[pt_ctr++] = pt_desc;
 
 	return pt_desc;
+}
+
+static inline void extract_va_metadata(u64 va, struct va_metadata *meta)
+{
+	u64 mask = 0xFFF;
+
+	meta->offset = va & mask;	//offset is calculated with last
+					//12 bits
+
+	mask = BITM(9);		//9 bits bitmask
+
+	meta->index[0] = (va >> 30) & mask;
+	meta->index[1] = (va >> 21) & mask;
+	meta->index[2] = (va >> 12) & mask;
+}
+
+static inline u64 pte_is_empty(u64 pte)
+{
+	return !(pte);
+}
+
+static u64 extract_addr(u64 pte)
+{
+	u64 addr = 0;
+
+	addr = pte & ~(0xF0000000000FFF);
+	if (addr < 0xFFFFFFFF)
+		return addr;
+
+	return KERNEL_START + (addr & 0xFFFFFFFF);
+}
+
+void map_pa_to_va_pg(u64 pa, u64 va, struct page_table_desc *pt_top,
+		     u64 flags)
+{
+	struct page_table_desc *pt_desc = pt_top;
+	volatile u64 *pt;
+	volatile u64 pte;
+	struct va_metadata meta;
+	u64 pte_flags = flags;
+	u16 pt_index;
+	u8 level, i;
+
+	extract_va_metadata(va, &meta);
+	meta.offset += pa;	//we want physical address at the end
+
+	if (!flags)
+		pte_flags = PTE_FLAGS_KERNEL_GENERIC;
+
+	for (i = pt_desc->level; i <= PT_LVL_MAX; i++) {
+		level = pt_desc->level;
+		pt = pt_desc->pt;
+		pt_index = meta.index[level - 1];
+
+		//L3
+		if (i == PT_LVL_MAX) {
+			pt[pt_index] = pt_entry(meta.offset, pte_flags);
+			break;
+		}
+
+		pte = pt[pt_index];
+		if (pte_is_empty(pte)) {
+			struct page_table_desc *new_pt_desc;
+
+			//we create next level
+			new_pt_desc = create_pt(0, level + 1);
+			pte = pt_entry((u64)new_pt_desc->pt, pte_flags);
+			place_pt_entry(pt_desc, pte, pt_index);
+			pt_desc = new_pt_desc;
+		} else {
+			//extract next table
+			pt_desc = (struct page_table_desc *)
+				(extract_addr(pte) + 0x1000);
+		}
+	}
+}
+
+void create_id_mapping(u64 start_paddr, u64 end_paddr,
+		       u64 pt, u64 flags)
+{
+	//for now lets assume T0/1_SZ is constant at 25, so we
+	//only care about 3 levels
+	struct page_table_desc *pt_desc;
+
+	pt_desc = create_pt(pt, 1);
+	int i;
+
+	for (i = start_paddr; i < end_paddr; i += PAGE_SIZE)
+		map_pa_to_va_pg(i, i, pt_desc, flags);
 }
 
 //traverse the page table and validate this vaddr range
